FossilOrigin-Name: 4d65f45b31c239d2a400376b30c486eb53113b0e7b594f2b7592df6027507e20

+++ src/whereInt.h
@@ -41,7 +41,7 @@ typedef struct WhereRightJoin WhereRightJoin;
 */
 struct WhereMemBlock {
   WhereMemBlock *pNext;      /* Next block in the chain */
-  u8 sz;                     /* Bytes of space */
+  u64 sz;                    /* Bytes of space */
 };
 
 /*
